pyproject.toml
[project]
name = "fedbrrrr"
version = "0.1.0"
description = "Economic and household financial health indicators dashboard using FRED data"
requires-python = ">=3.11"
dependencies = [
    "fredapi>=0.5.2",
    "python-dotenv>=1.0.0",
    "yfinance>=0.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]
